Module aws_sdk_macie::model

Modules

classification_type

See ClassificationType

classification_type_update

See ClassificationTypeUpdate

failed_s3_resource

See FailedS3Resource

member_account

See MemberAccount

s3_resource

See S3Resource

s3_resource_classification

See S3ResourceClassification

s3_resource_classification_update

See S3ResourceClassificationUpdate

Structs

ClassificationType

The classification type that Amazon Macie Classic applies to the associated S3

ClassificationTypeUpdate

The classification type that Amazon Macie Classic applies to the associated S3

FailedS3Resource

Includes details about the failed S3 resources.

MemberAccount

Contains information about the Amazon Macie Classic member account.

S3Resource

Contains information about the S3 resource. This data type is used as a request

S3ResourceClassification

The S3 resources that you want to associate with Amazon Macie Classic for monitoring

S3ResourceClassificationUpdate

The S3 resources whose classification types you want to update. This data type is used

Enums

S3ContinuousClassificationType

S3OneTimeClassificationType